adicionando links na controller do cliente e do atendimento

+++ src/main/java/fiap/_tdspr/evotech/gateways/controllers/ClienteController.java
@@ -3,11 +3,17 @@ package fiap._tdspr.evotech.gateways.controllers;
 import fiap._tdspr.evotech.domains.Cliente;
 import fiap._tdspr.evotech.usecases.ClienteService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.hateoas.EntityModel;
+import org.springframework.hateoas.CollectionModel;
+import org.springframework.hateoas.server.mvc.WebMvcLinkBuilder;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
+
+import static org.springframework.hateoas.server.mvc.WebMvcLinkBuilder.*;
 
 @RestController
 @RequestMapping("/clientes")
@@ -17,37 +23,61 @@ public class ClienteController {
     private ClienteService clienteService;
 
     @GetMapping
-    public ResponseEntity<List<Cliente>> getAllClientes() {
-        List<Cliente> clientes = clienteService.findAll();
-        return ResponseEntity.ok(clientes);
+    public ResponseEntity<CollectionModel<EntityModel<Cliente>>> getAllClientes() {
+        List<EntityModel<Cliente>> clientes = clienteService.findAll().stream()
+                .map(cliente -> EntityModel.of(cliente,
+                        linkTo(methodOn(ClienteController.class).getClienteById(cliente.getIdCliente())).withSelfRel(),
+                        linkTo(methodOn(ClienteController.class).getAllClientes()).withRel("allClientes")))
+                .collect(Collectors.toList());
+
+        return ResponseEntity.ok(CollectionModel.of(clientes,
+                linkTo(methodOn(ClienteController.class).getAllClientes()).withSelfRel()));
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Cliente> getClienteById(@PathVariable String id) {
+    public ResponseEntity<EntityModel<Cliente>> getClienteById(@PathVariable String id) {
         Optional<Cliente> cliente = clienteService.findById(id);
-        return cliente.map(ResponseEntity::ok)
-                .orElseGet(() -> ResponseEntity.notFound().build());
+        if (cliente.isPresent()) {
+            EntityModel<Cliente> clienteModel = EntityModel.of(cliente.get(),
+                    linkTo(methodOn(ClienteController.class).getClienteById(id)).withSelfRel(),
+                    linkTo(methodOn(ClienteController.class).getAllClientes()).withRel("allClientes"));
+
+            return ResponseEntity.ok(clienteModel);
+        }
+        return ResponseEntity.notFound().build();
     }
 
     @PostMapping
-    public ResponseEntity<Cliente> createCliente(@RequestBody Cliente cliente) {
+    public ResponseEntity<EntityModel<Cliente>> createCliente(@RequestBody Cliente cliente) {
         Cliente createdCliente = clienteService.save(cliente);
-        return ResponseEntity.status(201).body(createdCliente);
+        EntityModel<Cliente> clienteModel = EntityModel.of(createdCliente,
+                linkTo(methodOn(ClienteController.class).getClienteById(createdCliente.getIdCliente())).withSelfRel(),
+                linkTo(methodOn(ClienteController.class).getAllClientes()).withRel("allClientes"));
+
+        return ResponseEntity.status(201).body(clienteModel);
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Cliente> updateCliente(@PathVariable String id, @RequestBody Cliente cliente) {
+    public ResponseEntity<EntityModel<Cliente>> updateCliente(@PathVariable String id, @RequestBody Cliente cliente) {
         if (!clienteService.findById(id).isPresent()) {
             return ResponseEntity.notFound().build();
         }
         cliente.setIdCliente(id);
         Cliente updatedCliente = clienteService.save(cliente);
-        return ResponseEntity.ok(updatedCliente);
+
+        EntityModel<Cliente> clienteModel = EntityModel.of(updatedCliente,
+                linkTo(methodOn(ClienteController.class).getClienteById(id)).withSelfRel(),
+                linkTo(methodOn(ClienteController.class).getAllClientes()).withRel("allClientes"));
+
+        return ResponseEntity.ok(clienteModel);
     }
 
     @DeleteMapping("/{id}")
     public ResponseEntity<Void> deleteCliente(@PathVariable String id) {
-        clienteService.deleteById(id);
-        return ResponseEntity.noContent().build();
+        if (clienteService.findById(id).isPresent()) {
+            clienteService.deleteById(id);
+            return ResponseEntity.noContent().build();
+        }
+        return ResponseEntity.notFound().build();
     }
 }
